move clustering function to gammapy.utils.cluster
